native navigation

+++ Spark.API/DTOs/RegisterDto.cs
@@ -6,7 +6,7 @@
         public String LastName { get; set; }
         public String Email { get; set; }
         public String Password { get; set; }
-        public short Age { get; set; }
+        public int Age { get; set; }
         public String Gender { get; set; }
         public String Phone { get; set; }
     }
